The GummyJ language with the support for composite gummy modules
The GummyJ language is an extension to Java to support modular
event-driven programming based on the concept of gummy modules
(http://somayehmalakuti.com/event-based-modularization/).

In this distribution:

- gmprocess is an executable file that compile GummyJ programs. This has
been tested under Linux and Windows (gmprocess.exe)

- lib-core.jar is  the jar file of GummyJ runtime

-  file.gm is an example code in GummyJ, which must be compiled by
gmprocess. This results in a set of Java classes, which can be used to
execute the example. "lib-generated. jar" contains generated files for
our example.

-Tester.java is an example tester program in Java, which makes use of
these generated files.

-lib-test.jar and app-tester.jar are helper jar files.